titanicapi/requirements.txt
Django==3.1.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
urllib3==1.26.8
awswrangler==2.14.0